+++ FrontEnd/ndgradgate/package.json
@@ -32,6 +32,7 @@
 		"react-dom": "^18.2.0",
 		"react-router-dom": "^6.17.0",
 		"react-scripts": "^5.0.1",
+		"react-toastify": "^9.1.3",
 		"styled-components": "^6.1.0",
 		"typescript": "4.9.5",
 		"web-vitals": "^3.5.0"
