+++ FrontEnd/ndgradgate/src/Helpers/TabPanelHelper.tsx
@@ -1,0 +1,27 @@
+import { styled } from "styled-components";
+
+const TabPanelContainer = styled.div`
+  background: white;
+  padding: 20px;
+  border-radius: 10px;
+  margin-bottom: 20px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.05);
+`;
+export function TabPanel(props: {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}) {
+  const { children, value, index, ...other } = props;
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      {...other}
+    >
+      {value === index && <TabPanelContainer>{children}</TabPanelContainer>}
+    </div>
+  );
+}
